eve7: move future GeomDraw control in separate file
