+++ etc/eve7/geom.html
@@ -32,6 +32,8 @@
         function InitUI(handle) {
 
            jQuery.sap.registerModulePath("eve", evedir);
+           
+           console.log('ADD eve dir', evedir);
 
            new JSROOT.sap.ui.xmlview({
                id: "TopEveId",
